move db settings to db_store
