+++ db_store/src/lib.rs
@@ -1,4 +1,9 @@
 use std::str::FromStr;
+mod error;
+mod settings;
+
+pub use error::{Error, Result};
+pub use settings::Settings;
 
 pub mod meta;
 
@@ -6,16 +11,6 @@ pub mod meta;
 pub struct MetaValue<T> {
     key: String,
     value: T,
-}
-
-#[derive(thiserror::Error, Debug)]
-pub enum MetaError {
-    #[error("Sql error")]
-    SqlError(#[from] sqlx::Error),
-    #[error("Failed to decode meta value")]
-    DecodeError,
-    #[error("meta key not found {0}")]
-    NotFound(String),
 }
 
 impl<T> MetaValue<T> {
@@ -39,7 +34,7 @@ impl<T> MetaValue<T>
 where
     T: ToString,
 {
-    pub async fn insert<'c, E>(&self, exec: E) -> Result<(), MetaError>
+    pub async fn insert<'c, E>(&self, exec: E) -> Result
     where
         E: sqlx::Executor<'c, Database = sqlx::Postgres>,
     {
@@ -55,21 +50,21 @@ where
         exec: E,
         key: &str,
         default_fn: impl FnOnce() -> T,
-    ) -> Result<Self, MetaError>
+    ) -> Result<Self>
     where
         E: sqlx::Executor<'c, Database = sqlx::Postgres> + Copy,
     {
-        let result: Result<String, MetaError> = meta::fetch::<String>(exec, key).await;
+        let result: Result<String> = meta::fetch::<String>(exec, key).await;
 
         match result {
             Ok(str_val) => {
-                let value = str_val.parse().map_err(|_| MetaError::DecodeError)?;
+                let value = str_val.parse().map_err(|_| Error::DecodeError)?;
                 Ok(Self {
                     key: key.to_string(),
                     value,
                 })
             }
-            Err(MetaError::NotFound(_)) => {
+            Err(Error::NotFound(_)) => {
                 let value = default_fn();
                 let res = Self::new(key, value);
                 res.insert(exec).await?;
@@ -79,7 +74,7 @@ where
         }
     }
 
-    pub async fn update<'c, E>(&mut self, exec: E, new_val: T) -> Result<T, MetaError>
+    pub async fn update<'c, E>(&mut self, exec: E, new_val: T) -> Result<T>
     where
         E: sqlx::PgExecutor<'c>,
     {
