+++ mobile_index/src/error.rs
@@ -28,16 +28,14 @@ pub enum Error {
     NotFound(String),
     #[error("transaction error")]
     TransactionError(String),
-    #[error("meta error")]
-    MetaError(#[from] db_store::MetaError),
+    #[error("db error")]
+    DbError(#[from] db_store::Error),
 }
 
 #[derive(Error, Debug)]
 pub enum DecodeError {
     #[error("prost error")]
     Prost(#[from] helium_proto::DecodeError),
-    #[error("uri error")]
-    Uri(#[from] http::uri::InvalidUri),
     #[error("parse int error")]
     ParseInt(#[from] std::num::ParseIntError),
     #[error("datetime error")]
@@ -83,7 +81,6 @@ from_err!(EncodeError, prost::EncodeError);
 from_err!(EncodeError, serde_json::Error);
 
 // Decode Errors
-from_err!(DecodeError, http::uri::InvalidUri);
 from_err!(DecodeError, prost::DecodeError);
 from_err!(DecodeError, chrono::ParseError);
 from_err!(DecodeError, base64::DecodeError);
