+++ mobile_index/src/indexer.rs
@@ -19,7 +19,7 @@ pub struct Indexer {
 
 impl Indexer {
     pub async fn new(settings: &Settings) -> Result<Self> {
-        let pool = settings.database.connect().await?;
+        let pool = settings.database.connect(10).await?;
         Ok(Self {
             interval: settings.interval(),
             verifier_store: FileStore::from_settings(&settings.verifier).await?,
