+++ mobile_rewards/Cargo.toml
@@ -38,7 +38,6 @@ helium-crypto = { workspace = true }
 rust_decimal = {workspace = true}
 rust_decimal_macros = {workspace = true}
 tonic = {workspace = true}
-http = {workspace = true}
 rand = "*"
 async-trait = "*"
