+++ poc_iot_injector/Cargo.toml
@@ -24,7 +24,6 @@ chrono = {workspace = true}
 sqlx = {workspace = true}
 tonic = {workspace = true}
 tokio = {workspace = true}
-http = {workspace = true}
 triggered = {workspace = true}
 base64 = {workspace = true}
 sha2 = {workspace = true}
