+++ poc_iot_injector/src/settings.rs
@@ -1,8 +1,6 @@
 use crate::{Error, Result};
 use config::{Config, Environment, File};
-use http::Uri;
 use serde::Deserialize;
-use sqlx::{postgres::PgPoolOptions, Pool, Postgres};
 use std::{path::Path, time::Duration};
 
 #[derive(Debug, Deserialize)]
@@ -20,7 +18,7 @@ pub struct Settings {
     /// unix epoch)
     #[serde(default = "default_last_poc_submission")]
     pub last_poc_submission: i64,
-    pub database: Database,
+    pub database: db_store::Settings,
     pub transactions: node_follower::Settings,
     pub verifier: file_store::Settings,
     pub metrics: poc_metrics::Settings,
@@ -36,21 +34,6 @@ pub fn default_last_poc_submission() -> i64 {
 
 fn default_trigger_interval() -> u64 {
     1800
-}
-
-#[derive(Debug, Deserialize)]
-pub struct Database {
-    /// Max open connections to the database. (Default 10)
-    #[serde(default = "default_db_connections")]
-    pub max_connections: u32,
-    /// URL to access the postgres database. For example:
-    /// postgres://postgres:postgres@127.0.0.1:5432/mobile_index_db
-    #[serde(with = "http_serde::uri")]
-    pub url: Uri,
-}
-
-fn default_db_connections() -> u32 {
-    10
 }
 
 impl Settings {
@@ -84,15 +67,5 @@ impl Settings {
 
     pub fn trigger_interval(&self) -> Duration {
         Duration::from_secs(self.trigger)
-    }
-}
-
-impl Database {
-    pub async fn connect(&self) -> Result<Pool<Postgres>> {
-        let pool = PgPoolOptions::new()
-            .max_connections(self.max_connections)
-            .connect(&self.url.to_string())
-            .await?;
-        Ok(pool)
     }
 }
