+++ poc_iot_verifier/src/settings.rs
@@ -1,8 +1,6 @@
 use crate::{Error, Result};
 use config::{Config, Environment, File};
-use http::Uri;
 use serde::Deserialize;
-use sqlx::{postgres::PgPoolOptions, Pool, Postgres};
 use std::path::Path;
 
 #[derive(Debug, Deserialize, Clone)]
@@ -13,7 +11,7 @@ pub struct Settings {
     pub log: String,
     /// Cache location for generated verified reports
     pub cache: String,
-    pub database: Database,
+    pub database: db_store::Settings,
     pub follower: node_follower::Settings,
     pub ingest: file_store::Settings,
     pub entropy: file_store::Settings,
@@ -23,17 +21,6 @@ pub struct Settings {
 
 pub fn default_log() -> String {
     "poc_iot_verifier=debug,poc_store=info".to_string()
-}
-
-#[derive(Debug, Deserialize, Clone)]
-pub struct Database {
-    /// Max open connections to the database. If absent a default is calculated
-    /// by application code
-    pub max_connections: Option<u32>,
-    /// URL to access the postgres database. For example:
-    /// postgres://postgres:postgres@127.0.0.1:5432/mobile_index_db
-    #[serde(with = "http_serde::uri")]
-    pub url: Uri,
 }
 
 impl Settings {
@@ -58,18 +45,5 @@ impl Settings {
             .build()
             .and_then(|config| config.try_deserialize())
             .map_err(Error::from)
-    }
-}
-
-impl Database {
-    pub async fn connect(&self, default_max_connections: usize) -> Result<Pool<Postgres>> {
-        let pool = PgPoolOptions::new()
-            .max_connections(
-                self.max_connections
-                    .unwrap_or(default_max_connections as u32),
-            )
-            .connect(&self.url.to_string())
-            .await?;
-        Ok(pool)
     }
 }
